backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
huggingface_hub
langchain-huggingface
PyPDF2
faiss-cpu
sentence-transformers
sqlalchemy 
pymysql
python-dotenv